move commends from 'simpleTerminall' to 'commends'
